+++ skforecast/ForecasterSarimax/tests/test_fit.py
@@ -4,7 +4,6 @@ import re
 import pytest
 import numpy as np
 import pandas as pd
-from pmdarima.arima import ARIMA
 from skforecast.Sarimax import Sarimax
 from skforecast.ForecasterSarimax import ForecasterSarimax
 
@@ -79,20 +78,16 @@ def test_fit_last_window_stored(store_last_window):
         assert forecaster.last_window_ is None
 
 
-@pytest.mark.parametrize("regressor", 
-                         [ARIMA(order=(1, 0, 0)), 
-                          Sarimax(order=(1, 0, 0))], 
-                         ids = lambda reg: f'regressor: {type(reg)}')
 @pytest.mark.parametrize("y          , idx", 
                          [(y         , pd.RangeIndex(start=0, stop=50)), 
                           (y_datetime, pd.date_range(start='2000', periods=50, freq='YE'))], 
                          ids = lambda values: f'y, index: {type(values)}')
-def test_fit_extended_index_stored(regressor, y, idx):
+def test_fit_extended_index_stored(y, idx):
     """
     Test that values of self.regressor.arima_res_.fittedvalues.index are 
     stored after fitting in forecaster.extended_index_.
     """
-    forecaster = ForecasterSarimax(regressor=regressor)
+    forecaster = ForecasterSarimax(regressor=Sarimax(order=(1, 0, 0)))
     forecaster.fit(y=y)
 
     pd.testing.assert_index_equal(forecaster.extended_index_, idx)