+++ skforecast/preprocessing/preprocessing.py
@@ -11,6 +11,7 @@ import numpy as np
 import pandas as pd
 from sklearn.base import BaseEstimator
 from sklearn.base import TransformerMixin
+from numba import njit
 
 
 def _check_X_numpy_ndarray_1d(ensure_1d=True):
@@ -561,6 +562,62 @@ class DateTimeFeatureTransformer(BaseEstimator, TransformerMixin):
         return X_new
 
 
+@njit
+def _np_mean_jit(x):
+    return np.mean(x)
+
+
+@njit
+def _np_median_jit(x):
+    return np.median(x)
+
+
+@njit
+def _np_std_jit(x, ddof=1):
+    a_a, b_b = 0, 0
+    for i in x:
+        a_a = a_a + i
+        b_b = b_b + i * i
+    var = b_b / (len(x)) - ((a_a / (len(x))) ** 2)
+    var = var * (len(x) / (len(x) - ddof))
+    std = np.sqrt(var)
+
+    return std
+
+
+@njit
+def _np_sum_jit(x):
+    return np.sum(x)
+
+
+@njit
+def _np_min_jit(x):
+    return np.min(x)
+
+
+@njit
+def _np_max_jit(x):
+    return np.max(x)
+
+
+@njit
+def _np_min_max_ratio_jit(x):
+    return np.min(x) / np.max(x)
+
+
+@njit
+def _np_cv_jit(x):
+    a_a, b_b = 0, 0
+    for i in x:
+        a_a = a_a + i
+        b_b = b_b + i * i
+    var = b_b / (len(x)) - ((a_a / (len(x))) ** 2)
+    var = var * (len(x) / (len(x) - 1))
+    std = np.sqrt(var)
+
+    return std / np.mean(x)
+
+
 class RollingFeatures():
     """
     This class computes rolling features.
@@ -586,14 +643,10 @@ class RollingFeatures():
         - If 'right', the first point in the window is excluded from calculations.
         - If 'both', the no points in the window are excluded from calculations.
         - If 'neither', the first and last points in the window are excluded from calculations.
-    fill_strategy : str, list, default `None`
+    fill_strategy : str, float, default `None`
         Strategy to fill missing values in `transform_batch` method. Common pandas 
         methods for filling  missing values (e.g., 'ffill', 'bfill', 'mean', 
         'median', 'zero').
-    fill_strategy_predict : float, list, default `None`
-        Value(s) to fill missing data in the `transform` method. Can be a single 
-        `float` or a `list`. Filling with a constant value is appropriate since only 
-        one window is calculated in `transform`.
     
     Attributes
     ----------
@@ -609,8 +662,6 @@ class RollingFeatures():
         Closing method for the rolling window.
     fill_strategy : list
         Strategy to fill missing values in `transform_batch` method.
-    fill_strategy_predict : list
-        Value(s) to fill missing data in the `transform` method.
         
     """
 
@@ -621,8 +672,7 @@ class RollingFeatures():
         features_names: Union[str, list] = None, 
         min_periods: Union[int, list] = None,
         closed: Union[str, list] = 'left',
-        fill_strategy: Union[str, list] = None,
-        fill_strategy_predict: Union[float, list] = None
+        fill_strategy: Union[str, float] = None
     ) -> None:
         
         # TODO: Return of method or method to set attributes?
@@ -631,9 +681,7 @@ class RollingFeatures():
             window_sizes, 
             features_names, 
             min_periods,
-            closed,
-            fill_strategy,
-            fill_strategy_predict
+            fill_strategy
         )
 
         # (
@@ -642,44 +690,42 @@ class RollingFeatures():
         #     features_names,
         #     min_periods,
         #     closed,
-        #     fill_strategy,
-        #     fill_strategy_predict
+        #     fill_strategy
         # ) = self._validate_preprocess_params(
         #         stats, 
         #         window_sizes, 
         #         features_names, 
         #         min_periods,
         #         closed,
-        #         fill_strategy,
-        #         fill_strategy_predict
+        #         fill_strategy
         #     )
         
         # self.stats = stats
         # self.window_sizes = window_sizes
+        # self.max_window_size = max(window_sizes)
         # self.features_names = features_names
         # self.min_periods = min_periods
         # self.closed = closed
         # self.fill_strategy = fill_strategy
-        # self.fill_strategy_predict = fill_strategy_predict
 
         window_params_list = []
         for i in range(len(self.stats)):
-            window_params = (self.window_sizes[i], self.min_periods[i], self.closed[i])
+            window_params = (self.window_sizes[i], self.min_periods[i])
             window_params_list.append(window_params)
 
         # Find unique window parameter combinations
-        _unique_rolling_windows = {}
+        unique_rolling_windows = {}
         for i, params in enumerate(window_params_list):
-            key = f"{params[0]}_{params[1]}_{params[2]}"
-            if key not in _unique_rolling_windows:
-                _unique_rolling_windows[key] = {
-                    'params': {'window': params[0], 'min_periods': params[1], 'closed': params[2]},
+            key = f"{params[0]}_{params[1]}"
+            if key not in unique_rolling_windows:
+                unique_rolling_windows[key] = {
+                    'params': {'window': params[0], 'min_periods': params[1]},
                     'stats_idx': [], 
                     'rolling_obj': None
                 }
-            _unique_rolling_windows[key]['stats_idx'].append(i)
+            unique_rolling_windows[key]['stats_idx'].append(i)
 
-        self._unique_rolling_windows = _unique_rolling_windows
+        self.unique_rolling_windows = unique_rolling_windows
 
 
     def _validate_preprocess_params(
@@ -705,7 +751,8 @@ class RollingFeatures():
                 f"`stats` must be a string or a list of strings. Got {type(stats)}."
             )
         
-        allowed_stats = ['mean', 'std', 'min', 'max', 'sum', 'median', 'var']
+        # TODO: Review list with pandas docs
+        allowed_stats = ['mean', 'std', 'min', 'max', 'sum', 'median', 'ratio_min_max', 'coef_variation']
         for stat in set(stats):
             if stat not in allowed_stats:
                 raise ValueError(
@@ -714,6 +761,7 @@ class RollingFeatures():
         
         self.stats = stats
         n_stats = len(stats)
+        self.n_stats = len(stats)
         
         # window_sizes
         if isinstance(window_sizes, int):
@@ -730,6 +778,8 @@ class RollingFeatures():
             raise TypeError(
                 f"`window_sizes` must be an int or a list of ints. Got {type(window_sizes)}."
             )
+        
+        self.max_window_size = max(window_sizes)
         
         # features_names
         if features_names is None:
@@ -766,29 +816,29 @@ class RollingFeatures():
             )
         
         # closed
-        allowed_closed = ['right', 'left', 'both', 'neither']
-        if isinstance(closed, str):
-            closed = [closed] * n_stats
-        elif isinstance(closed, list):
-            n_closed = len(closed)
-            if n_closed != n_stats:
-                raise ValueError(
-                    (f"Length of `closed` list ({n_closed}) "
-                     f"must match length of `stats` list ({n_stats}).")
-                )
-        else:
-            raise TypeError(
-                f"`closed` must be a string or a list of strings. Got {type(window_sizes)}."
-            )
+        # allowed_closed = ['right', 'left', 'both', 'neither']
+        # if isinstance(closed, str):
+        #     closed = [closed] * n_stats
+        # elif isinstance(closed, list):
+        #     n_closed = len(closed)
+        #     if n_closed != n_stats:
+        #         raise ValueError(
+        #             (f"Length of `closed` list ({n_closed}) "
+        #              f"must match length of `stats` list ({n_stats}).")
+        #         )
+        # else:
+        #     raise TypeError(
+        #         f"`closed` must be a string or a list of strings. Got {type(window_sizes)}."
+        #     )
         
-        allowed_closed = ['right', 'left', 'both', 'neither']
-        for c in set(closed):
-            if c not in allowed_closed:
-                raise ValueError(
-                    f"'{c}' is not allowed. Allowed `closed` values are {allowed_closed}."
-                )
+        # allowed_closed = ['right', 'left', 'both', 'neither']
+        # for c in set(closed):
+        #     if c not in allowed_closed:
+        #         raise ValueError(
+        #             f"'{c}' is not allowed. Allowed `closed` values are {allowed_closed}."
+        #         )
         
-        self.closed = closed
+        # self.closed = closed
         
         # fill_strategy
         if fill_strategy is None:
@@ -834,22 +884,34 @@ class RollingFeatures():
             raise TypeError(
                 f"`fill_strategy_predict` must be a float or a list of float. Got {type(fill_strategy_predict)}"
             )
-        
-        # TODO: Complete
-        # allowed_fill_strategy_predict = ['right', 'left', 'both', 'neither']
-        # for fs in set(fill_strategy_predict):
-        #     if fs not in allowed_fill_strategy_predict:
-        #         raise ValueError(
-        #             (f"'{fs}' is not allowed. Allowed `fill_strategy_predict` "
-        #              f"values are {allowed_fill_strategy_predict}.")
-        #         )
-        
-        self.fill_strategy_predict = fill_strategy_predict
+
+
+    def _apply_stat_pandas(self, rolling_obj, stat):
+        """
+        """
+        if stat == 'mean':
+            return rolling_obj.mean()
+        elif stat == 'median':
+            return rolling_obj.median()
+        elif stat == 'std':
+            return rolling_obj.std()
+        elif stat == 'sum':
+            return rolling_obj.sum()
+        elif stat == 'min':
+            return rolling_obj.min()
+        elif stat == 'max':
+            return rolling_obj.max()
+        elif stat == 'ratio_min_max':
+            return rolling_obj.min() / rolling_obj.max()
+        elif stat == 'coef_variation':
+            return rolling_obj.std() / rolling_obj.mean()
+        else:
+            raise ValueError(f"Statistic '{stat}' is not implemented")
 
 
     def transform_batch(
         self, 
-        series: pd.Series
+        X: pd.Series
     ) -> pd.DataFrame:
         """
         Transform an entire pandas Series using rolling windows and compute the 
@@ -857,7 +919,7 @@ class RollingFeatures():
 
         Parameters
         ----------
-        series : pandas Series
+        X : pandas Series
             The input data series to transform.
 
         Returns
@@ -867,131 +929,209 @@ class RollingFeatures():
         
         """
         
-        results = pd.DataFrame(index=series.index)
-        series_name = series.name if series.name is not None else 'y'
+        rolling_features = pd.DataFrame(index=X.index)
 
-        for k in self._unique_rolling_windows.keys():
-            print(self._unique_rolling_windows[k]['params'])
-            rolling_obj = series.rolling(**self._unique_rolling_windows[k]['params'])
-            self._unique_rolling_windows[k]['rolling_obj'] = rolling_obj
+        for k in self.unique_rolling_windows.keys():
+            rolling_obj = X.rolling(**self.unique_rolling_windows[k]['params'])
+            self.unique_rolling_windows[k]['rolling_obj'] = rolling_obj
         
         for i, stat in enumerate(self.stats):
             window_size = self.window_sizes[i]
             min_periods = self.min_periods[i]
-            closed = self.closed[i]
             fill_strategy = self.fill_strategy[i]
 
-            key = f"{window_size}_{min_periods}_{closed}"
-            rolling_obj = self._unique_rolling_windows[key]['rolling_obj']
+            key = f"{window_size}_{min_periods}"
+            rolling_obj = self.unique_rolling_windows[key]['rolling_obj']
             
-            # Generate feature name
             if self.features_names is not None:
                 feature_name = self.features_names[i]
             else:
-                feature_name = f"_{series_name}_{stat}"
+                feature_name = f"roll_{stat}_{window_size}"  # roll_mean_7
             
             # Get the function corresponding to the statistic
-            if stat == 'mean':
-                stat_series = rolling_obj.mean()
-            elif stat == 'std':
-                stat_series = rolling_obj.std()
-            elif stat == 'var':
-                stat_series = rolling_obj.var()
-            elif stat == 'min':
-                stat_series = rolling_obj.min()
-            elif stat == 'max':
-                stat_series = rolling_obj.max()
-            elif stat == 'sum':
-                stat_series = rolling_obj.sum()
-            elif stat == 'median':
-                stat_series = rolling_obj.median()
+            stat_series = self._apply_stat_pandas(rolling_obj=rolling_obj, stat=stat)
+
+
+            # else:
+            #     # Try to convert fill_strategy to a number
+            #     try:
+            #         fill_value = float(fill_strategy)
+            #         stat_series = stat_series.fillna(fill_value)
+            #     except ValueError:
+            #         raise ValueError(f"fill_strategy '{fill_strategy}' is not recognized")
+            
+            rolling_features[feature_name] = stat_series
+
+        rolling_features = rolling_features.iloc[self.max_window_size - 1:]
+
+        # TODO: Review
+        # Fill missing values if fill_strategy is specified
+        if fill_strategy is not None:
+            if fill_strategy == 'mean':
+                rolling_features = rolling_features.fillna(rolling_features.mean())
+            elif fill_strategy == 'median':
+                rolling_features = rolling_features.fillna(rolling_features.median())
+            elif fill_strategy == 'ffill':
+                rolling_features = rolling_features.ffill()
+            elif fill_strategy == 'bfill':
+                rolling_features = rolling_features.bfill()
+            # elif fill_strategy == 'zero':
+            #     rolling_features = rolling_features.fillna(0)
             else:
-                raise ValueError(f"Statistic '{stat}' is not implemented")
-            
-            # TODO: Complete
-            # Fill missing values if fill_strategy is specified
-            if fill_strategy:
-                if fill_strategy == 'mean':
-                    stat_series = stat_series.fillna(stat_series.mean())
-                elif fill_strategy == 'median':
-                    stat_series = stat_series.fillna(stat_series.median())
-                elif fill_strategy == 'ffill':
-                    stat_series = stat_series.fillna(method='ffill')
-                elif fill_strategy == 'bfill':
-                    stat_series = stat_series.fillna(method='bfill')
-                elif fill_strategy == 'zero':
-                    stat_series = stat_series.fillna(0)
-                else:
-                    # Try to convert fill_strategy to a number
-                    try:
-                        fill_value = float(fill_strategy)
-                        stat_series = stat_series.fillna(fill_value)
-                    except ValueError:
-                        raise ValueError(f"fill_strategy '{fill_strategy}' is not recognized")
-            
-            results[feature_name] = stat_series
+                rolling_features = rolling_features.fillna(fill_strategy)
         
-        return results
+        return rolling_features
+
+
+    def _apply_stat_numpy(self, X_window, stat):
+        """
+        """
+        if stat == 'mean':
+            return _np_mean_jit(X_window)
+        elif stat == 'median':
+            return _np_median_jit(X_window)
+        elif stat == 'std':
+            return _np_std_jit(X_window)
+        elif stat == 'sum':
+            return _np_sum_jit(X_window)
+        elif stat == 'min':
+            return _np_min_jit(X_window)
+        elif stat == 'max':
+            return _np_max_jit(X_window)
+        elif stat == 'ratio_min_max':
+            return _np_min_max_ratio_jit(X_window)
+        elif stat == 'coef_variation':
+            return _np_cv_jit(X_window)
+        else:
+            raise ValueError(f"Statistic '{stat}' is not implemented")
+
+
+    def transform_batch_2(
+        self, 
+        X: pd.Series
+    ) -> pd.DataFrame:
+        """
+        Transform an entire pandas Series using rolling windows and compute the 
+        specified statistics.
+
+        Parameters
+        ----------
+        X : pandas Series
+            The input data series to transform.
+
+        Returns
+        -------
+        results : pandas DataFrame
+            A DataFrame containing the rolling features.
+        
+        """
+
+        for k in self.unique_rolling_windows.keys():
+            rolling_obj = X.rolling(**self.unique_rolling_windows[k]['params'])
+            self.unique_rolling_windows[k]['rolling_obj'] = rolling_obj
+        
+        rolling_features = []
+        for i, stat in enumerate(self.stats):
+            window_size = self.window_sizes[i]
+            min_periods = self.min_periods[i]
+            fill_strategy = self.fill_strategy[i]
+
+            key = f"{window_size}_{min_periods}"
+            rolling_obj = self.unique_rolling_windows[key]['rolling_obj']
+            
+            if self.features_names is not None:
+                feature_name = self.features_names[i]
+            else:
+                feature_name = f"roll_{stat}_{window_size}"  # roll_mean_7
+            
+            # Get the function corresponding to the statistic
+            stat_series = self._apply_stat_pandas(rolling_obj=rolling_obj, stat=stat)
+            stat_series.name = feature_name
+
+
+            # else:
+            #     # Try to convert fill_strategy to a number
+            #     try:
+            #         fill_value = float(fill_strategy)
+            #         stat_series = stat_series.fillna(fill_value)
+            #     except ValueError:
+            #         raise ValueError(f"fill_strategy '{fill_strategy}' is not recognized")
+            
+            rolling_features.append(stat_series)
+
+        rolling_features = pd.concat(rolling_features, axis=1)
+        # rolling_features.columns = self.features_names
+        rolling_features = rolling_features.iloc[self.max_window_size - 1:]
+
+        # TODO: Review
+        # Fill missing values if fill_strategy is specified
+        if fill_strategy is not None:
+            if fill_strategy == 'mean':
+                rolling_features = rolling_features.fillna(rolling_features.mean())
+            elif fill_strategy == 'median':
+                rolling_features = rolling_features.fillna(rolling_features.median())
+            elif fill_strategy == 'ffill':
+                rolling_features = rolling_features.ffill()
+            elif fill_strategy == 'bfill':
+                rolling_features = rolling_features.bfill()
+            # elif fill_strategy == 'zero':
+            #     rolling_features = rolling_features.fillna(0)
+            else:
+                rolling_features = rolling_features.fillna(fill_strategy)
+        
+        return rolling_features
+
+
+    def _apply_stat_numpy(self, X_window, stat):
+        """
+        """
+        if stat == 'mean':
+            return _np_mean_jit(X_window)
+        elif stat == 'median':
+            return _np_median_jit(X_window)
+        elif stat == 'std':
+            return _np_std_jit(X_window)
+        elif stat == 'sum':
+            return _np_sum_jit(X_window)
+        elif stat == 'min':
+            return _np_min_jit(X_window)
+        elif stat == 'max':
+            return _np_max_jit(X_window)
+        elif stat == 'ratio_min_max':
+            return _np_min_max_ratio_jit(X_window)
+        elif stat == 'coef_variation':
+            return _np_cv_jit(X_window)
+        else:
+            raise ValueError(f"Statistic '{stat}' is not implemented")
 
 
     def transform(
         self, 
-        array: np.ndarray
+        X: np.ndarray
     ) -> np.ndarray:
         """
         Transform a numpy array using rolling windows and compute the specified statistics.
 
         Parameters
         ----------
-        array : np.ndarray
+        X : np.ndarray
             The input data array to transform. Should be at least as long as the maximum window size.
 
         Returns
         -------
-        results : numpy array
+        rolling_features : numpy array
             An array containing the computed statistics.
         
         """
         
-        results = []
-        
-        array = np.asarray(array)
-        n_samples = len(array)
-        max_window_size = max(self.window_sizes)
-        
-        if n_samples < max_window_size:
-            raise ValueError(f"Input array must have at least {max_window_size} observations")
-        
+        X = X[~np.isnan(X)]
+
+        rolling_features = np.full(shape=self.n_stats, fill_value=np.nan, dtype=float)
         for i, stat in enumerate(self.stats):
             window_size = self.window_sizes[i]
-            data_window = array[-window_size:]  # Take the last window_size observations
+            X_window = X[-window_size:]  # Take the last window_size observations
             
             # Compute the statistic using numpy, handling NaNs
-            if stat == 'mean':
-                value = np.nanmean(data_window)
-            elif stat == 'std':
-                value = np.nanstd(data_window, ddof=1)  # ddof=1 for sample std
-            elif stat == 'var':
-                value = np.nanvar(data_window, ddof=1)
-            elif stat == 'min':
-                value = np.nanmin(data_window)
-            elif stat == 'max':
-                value = np.nanmax(data_window)
-            elif stat == 'sum':
-                value = np.nansum(data_window)
-            elif stat == 'median':
-                value = np.nanmedian(data_window)
-            else:
-                raise ValueError(f"Statistic '{stat}' is not implemented")
-            
-            # Handle fill_strategy_predict
-            fill_value = self.fill_strategy_predict[i]
-            if np.isnan(value):
-                if fill_value is not None:
-                    value = fill_value
-                else:
-                    value = np.nan  # Leave as NaN if no fill_value is specified
-            
-            results.append(value)
+            rolling_features[i] = self._apply_stat_numpy(X_window, stat)
         
-        return np.array(results)
+        return rolling_features